implementacion de seguridad en los metodos que no la tenian

+++ src/main/java/com/example/zonadegolbackend/services/LigaService.java
@@ -2,9 +2,12 @@ package com.example.zonadegolbackend.services;
 
 
 import com.example.zonadegolbackend.entity.*;
+import com.example.zonadegolbackend.enums.Rol;
 import com.example.zonadegolbackend.repository.*;
 import jakarta.transaction.Transactional;
 import lombok.AllArgsConstructor;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -21,6 +24,7 @@ public class LigaService {
     private final TemporadaLigaRepository temporadaLigaRepository;
     private final ClasificacionService clasificacionService;
     private final TemporadaService temporadaService;
+    private final UsuarioRepository usuarioRepository;
 
     public List<Liga> findAll() {
         return ligaRepository.findAll();
@@ -46,6 +50,15 @@ public class LigaService {
 //    }
 
     public Liga crearLiga(Liga liga){
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        String username = authentication.getName();
+
+        Usuario usuarioAutenticado = usuarioRepository.findByUsername(username)
+                .orElseThrow(() -> new RuntimeException("Usuario no encontrado"));
+
+        if (usuarioAutenticado.getRol() != Rol.ADMIN) {
+            throw new RuntimeException("Solo un administrador crear una liga");
+        }
 
         Liga nuevaLiga = new Liga();
 
@@ -53,6 +66,7 @@ public class LigaService {
         nuevaLiga.setNumEquipos(liga.getNumEquipos());
         nuevaLiga.setDescripcion(liga.getDescripcion());
         nuevaLiga.setFecha_fundacion(liga.getFecha_fundacion());
+        nuevaLiga.setImagen(liga.getImagen());
         ligaRepository.save(nuevaLiga);
 
         Temporada temporadaActual = temporadaService.buscarTemporadaMasReciente();
@@ -72,6 +86,16 @@ public class LigaService {
     }
 
     public Liga editarLiga(Integer id, Liga liga){
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        String username = authentication.getName();
+
+        Usuario usuarioAutenticado = usuarioRepository.findByUsername(username)
+                .orElseThrow(() -> new RuntimeException("Usuario no encontrado"));
+
+        if (usuarioAutenticado.getRol() != Rol.ADMIN) {
+            throw new RuntimeException("Solo un administrador puede editar una liga");
+        }
+
         Liga ligaExistente = ligaRepository.findById(id).orElseThrow(() -> new RuntimeException("Liga no encontrada"));
         ligaExistente.setNombre(liga.getNombre());
         ligaExistente.setNumEquipos(liga.getNumEquipos());
@@ -82,6 +106,16 @@ public class LigaService {
     }
 
     public void eliminarLiga(Integer id){
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        String username = authentication.getName();
+
+        Usuario usuarioAutenticado = usuarioRepository.findByUsername(username)
+                .orElseThrow(() -> new RuntimeException("Usuario no encontrado"));
+
+        if (usuarioAutenticado.getRol() != Rol.ADMIN) {
+            throw new RuntimeException("Solo un administrador puede eliminar una liga");
+        }
+
         Liga ligaExistente = ligaRepository.findById(id).orElseThrow(() -> new RuntimeException("Liga no encontrada"));
 
         ligaRepository.delete(ligaExistente);
